// src/features/order/models/document-detail.model.ts
export interface IDocumentDetailItem {
  no: number;
  id: string;
  tenHang?: string;
  dvt?: string;
  soLuong?: number;
  donGia?: number;
  tien?: number;
  phanTramCk?: number;
  tienCk?: number;
  phanTramThue?: number;
  tienThue?: number;
  phanTramGiamTru?: number;
  tienGiamTru?: number;
  tienSauGiamTru?: number;
}